+++ src/Sql/Sql/ChangeLog.md
@@ -20,6 +20,7 @@
 ## Upcoming Release
 * Add Get/Set AzSqlDatabaseBackupShortTermRetentionPolicy
 * Fix issue where not being logged into Azure account would result in nullref exception when executing SQL cmdlets
+* Fixed null ref exception in Get-AzSqlCapability
 
 ## Version 1.1.0
 * Update incorrect online help URLs