+++ src/Sql/Sql/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Add support for SQL DB Hyperscale tier
 
 ## Version 1.2.0
 * Add Get/Set AzSqlDatabaseBackupShortTermRetentionPolicy